+++ test/array/average.spec.js
@@ -10,12 +10,12 @@ test('returns a number', t => {
 
 test('returns the average of the sum of all numerical items', t => {
     let input = [1, 2, 3];
-    let expectation = 3;
+    let expectation = 2;
     t.is(average(input), expectation);
 });
 
 test('returns the average of the sum of all numerical items even if here are negative items', t => {
     let input = [-1, 0, 1, 2, 3];
-    let expectation = 2.5;
+    let expectation = 1;
     t.is(average(input), expectation);
 });