+++ test/array/index.spec.js
@@ -1,88 +1,100 @@
-import * as Array from '../../src/array';
+import * as arr from '../../src/array';
 const test = require('ava');
 
-test('exports a function "(Array.)occurrenceMap"', t => {
-    t.true(typeof (Array.occurrenceMap) === 'function')
+test('exports a function "(arr.)occurrenceMap"', t => {
+    t.true(typeof (arr.occurrenceMap) === 'function')
 });
 
-test('exports a function "(Array.)isEmpty"', t => {
-    t.true(typeof (Array.isEmpty) === 'function')
+test('exports a function "(arr.)isEmpty"', t => {
+    t.true(typeof (arr.isEmpty) === 'function')
 });
 
-test('exports a function "(Array.)removeDuplicates"', t => {
-    t.true(typeof (Array.removeDuplicates) === 'function')
+test('exports a function "(arr.)removeDuplicates"', t => {
+    t.true(typeof (arr.removeDuplicates) === 'function')
 });
 
-test('exports a function "(Array.)shallowClone"', t => {
-    t.true(typeof (Array.shallowClone) === 'function')
+test('exports a function "(arr.)shallowClone"', t => {
+    t.true(typeof (arr.shallowClone) === 'function')
 });
 
-test('exports a function "(Array.)lowestNumber"', t => {
-    t.true(typeof (Array.lowestNumber) === 'function')
+test('exports a function "(arr.)lowestNumber"', t => {
+    t.true(typeof (arr.lowestNumber) === 'function')
 });
 
-test('exports a function "(Array.)highestNumber"', t => {
-    t.true(typeof (Array.highestNumber) === 'function')
+test('exports a function "(arr.)highestNumber"', t => {
+    t.true(typeof (arr.highestNumber) === 'function')
 });
-test('exports a function "(Array.)closestNumber"', t => {
-    t.true(typeof (Array.closestNumber) === 'function')
-});
-
-test('exports a function "(Array.)indexOfLowestNumber"', t => {
-    t.true(typeof (Array.indexOfLowestNumber) === 'function')
+test('exports a function "(arr.)closestNumber"', t => {
+    t.true(typeof (arr.closestNumber) === 'function')
 });
 
-test('exports a function "(Array.)indexOfHighestNumber"', t => {
-    t.true(typeof (Array.indexOfHighestNumber) === 'function')
+test('exports a function "(arr.)indexOfLowestNumber"', t => {
+    t.true(typeof (arr.indexOfLowestNumber) === 'function')
 });
 
-test('exports a function "(Array.)splitInHalf"', t => {
-    t.true(typeof (Array.splitInHalf) === 'function')
-});
-test('exports a function "(Array.)longestString"', t => {
-    t.true(typeof (Array.longestString) === 'function')
+test('exports a function "(arr.)indexOfHighestNumber"', t => {
+    t.true(typeof (arr.indexOfHighestNumber) === 'function')
 });
 
-test('exports a function "(Array.)shortestString"', t => {
-    t.true(typeof (Array.shortestString) === 'function')
+test('exports a function "(arr.)splitInHalf"', t => {
+    t.true(typeof (arr.splitInHalf) === 'function')
+});
+test('exports a function "(arr.)longestString"', t => {
+    t.true(typeof (arr.longestString) === 'function')
 });
 
-test('exports a function "(Array.)shortestStringLength"', t => {
-    t.true(typeof (Array.shortestStringLength) === 'function')
+test('exports a function "(arr.)shortestString"', t => {
+    t.true(typeof (arr.shortestString) === 'function')
 });
 
-test('exports a function "(Array.)longestStringLength"', t => {
-    t.true(typeof (Array.longestStringLength) === 'function')
+test('exports a function "(arr.)shortestStringLength"', t => {
+    t.true(typeof (arr.shortestStringLength) === 'function')
 });
 
-test('exports a function "(Array.)oddItems"', t => {
-    t.true(typeof (Array.oddItems) === 'function')
+test('exports a function "(arr.)longestStringLength"', t => {
+    t.true(typeof (arr.longestStringLength) === 'function')
 });
 
-test('exports a function "(Array.)evenItems"', t => {
-    t.true(typeof (Array.evenItems) === 'function')
+test('exports a function "(arr.)oddItems"', t => {
+    t.true(typeof (arr.oddItems) === 'function')
 });
 
-test('exports a function "(Array.)mergeArrayUnique"', t => {
-    t.true(typeof (Array.mergeArrayUnique) === 'function')
+test('exports a function "(arr.)evenItems"', t => {
+    t.true(typeof (arr.evenItems) === 'function')
 });
 
-test('exports a function "(Array.)mergeArray"', t => {
-    t.true(typeof (Array.mergeArray) === 'function')
+test('exports a function "(arr.)mergeArrayUnique"', t => {
+    t.true(typeof (arr.mergeArrayUnique) === 'function')
 });
 
-test('exports a function "(Array.)clean"', t => {
-    t.true(typeof (Array.clean) === 'function')
+test('exports a function "(arr.)mergeArray"', t => {
+    t.true(typeof (arr.mergeArray) === 'function')
 });
 
-test('exports a function "(Array.)cleanFalsy"', t => {
-    t.true(typeof (Array.cleanFalsy) === 'function')
+test('exports a function "(arr.)clean"', t => {
+    t.true(typeof (arr.clean) === 'function')
 });
 
-test('exports a function "(Array.)sum"', t => {
+test('exports a function "(arr.)cleanFalsy"', t => {
+    t.true(typeof (arr.cleanFalsy) === 'function')
+});
+
+test('exports a function "(arr.)sum"', t => {
     t.true(typeof (arr.sum) === 'function')
 });
 
-test('exports a function "(Array.)average"', t => {
+test('exports a function "(arr.)average"', t => {
     t.true(typeof (arr.average) === 'function')
+});
+
+test('exports a function "(arr.)nthItems"', t => {
+    t.true(typeof (arr.nthItems) === 'function')
+});
+
+test('exports a function "(arr.)sortAsc"', t => {
+    t.true(typeof (arr.sortAsc) === 'function')
+});
+
+test('exports a function "(arr.)sortDesc"', t => {
+    t.true(typeof (arr.sortDesc) === 'function')
 });